Create msu_cout.cpp
In the file "MSU_chars/main.cpp", write a program that characters standard input and prints the number of MSU characters in it. Your program should read input until the period character (.) is encountered.
For example, if the input is MSU ROCKS., your program should output The count of letters ("MSU") is 4., because the letters M, S, and U appear 4 times.